+++ common/src/main/java/net/mehvahdjukaar/polytone/slotify/SimpleSprite.java
@@ -6,6 +6,7 @@ import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.components.Renderable;
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.resources.ResourceLocation;
@@ -14,7 +15,7 @@ import org.joml.Matrix4f;
 import java.util.Optional;
 
 public record SimpleSprite(ResourceLocation texture, float x, float y, float width, float height, float z,
-                           Optional<String> tooltip) {//, Optional<ScreenSupplier> screenSupp) {
+                           Optional<String> tooltip) implements Renderable {//, Optional<ScreenSupplier> screenSupp) {
 
     public static final Codec<SimpleSprite> CODEC = RecordCodecBuilder.create(i -> i.group(
             ResourceLocation.CODEC.fieldOf("texture").forGetter(SimpleSprite::texture),
@@ -29,9 +30,10 @@ public record SimpleSprite(ResourceLocation texture, float x, float y, float wid
     ).apply(i, SimpleSprite::new));
 
 
-    public void render(GuiGraphics poseStack) {
+    @Override
+    public void render(GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick) {
         TextureAtlasSprite sprite = Minecraft.getInstance().getGuiSprites().getSprite(texture);
-        blit(poseStack.pose().last().pose(), sprite.atlasLocation(), x, x + width, y, y + height, z,
+        blit(guiGraphics.pose().last().pose(), sprite.atlasLocation(), x, x + width, y, y + height, z,
                 sprite.getU0(), sprite.getU1(), sprite.getV0(), sprite.getV1());
     }
 
@@ -40,7 +42,8 @@ public record SimpleSprite(ResourceLocation texture, float x, float y, float wid
                             float blitOffset, float minU, float maxU, float minV, float maxV) {
         RenderSystem.setShaderTexture(0, atlasLoc);
         RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        BufferBuilder bufferBuilder = Tesselator.getInstance().begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_TEX);;
+        BufferBuilder bufferBuilder = Tesselator.getInstance().begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_TEX);
+        ;
         bufferBuilder.addVertex(matrix, x1, y1, blitOffset).setUv(minU, minV);
         bufferBuilder.addVertex(matrix, x1, y2, blitOffset).setUv(minU, maxV);
         bufferBuilder.addVertex(matrix, x2, y2, blitOffset).setUv(maxU, maxV);