+++ common/src/main/java/net/mehvahdjukaar/polytone/slotify/SlotifyScreen.java
@@ -5,7 +5,7 @@ import net.minecraft.client.gui.GuiGraphics;
 
 public interface SlotifyScreen {
 
-    void polytone$renderExtraSprites(GuiGraphics poseStack);
+    void polytone$renderExtraSprites(GuiGraphics poseStack, int mouseX, int mouseY, float partialTicks);
 
     boolean polytone$hasSprites();